requirements.txt
Django~=3.2.8
djangorestframework~=3.12.4
PyJWT~=2.3.0
factory-boy~=3.2.0
Faker~=9.3.1
django-cors-headers~=3.10.0
drf-yasg~=1.20.0
pytz~=2021.3
locust~=2.4.1